# api/openapi.yaml
openapi: 3.0.3
info:
  title: Fish Landing + Dashboard API
  version: 0.1.0
  description: API contract for the Fish landing page, Ocean Network supply dashboard, provider pilot, and waitlist.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/meal/order:
    post:
      summary: Guest Fish meal-counter order
      description: Capped guest route for the public meal counter. Uses the same chat request shape without exposing a Fish API key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FishChatRequest'
      responses:
        '200':
          description: Meal order completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FishChatCompletion'
...
  /api/providers/pilot:
    get:
      summary: Public-safe provider pilot registry and selected allowlist
      parameters:
        - in: query
          name: providerId
          schema:
            type: string
          description: Optional provider id to check against the allowlist.
        - in: query
          name: workloadType
          schema:
            type: string
            default: chat_batch
        - in: query
          name: model
          schema:
            type: string
            default: fish-demo-chat
      responses:
        '200':
          description: Provider pilot registry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderPilotRegistry'
...
  /api/proof/capacity-settlements:
    get:
      summary: Public-safe FISH Capacity Pool settlement summary
      description: Lists operator-reported paid-demand settlement records. Records are snapshot data until onchain verification is configured.
      responses:
        '200':
          description: Capacity settlement summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacitySettlementSummary'
    post:
      summary: Admin-only FISH Capacity Pool settlement record
      description: Records that paid demand was settled into the FISH Capacity Pool. The route stores public-safe amount and transaction metadata only.
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapacitySettlementRequest'
      responses:
        '201':
          description: Capacity settlement recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  idempotent:
                    type: boolean
                  settlement:
                    $ref: '#/components/schemas/CapacitySettlement'
        '400':
          description: Invalid capacity settlement request
        '401':
          description: Admin token required
...
  /api/proof/receipts/export:
    get:
      summary: Admin-only provider receipt ledger export
      security:
        - adminToken: []
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [csv, json]
            default: csv
        - in: query
          name: provider
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [succeeded, failed, timed_out, not_allowed]
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: CSV or JSON receipt ledger export for operator review
          content:
            text/csv:
              schema:
                type: string
            application/json:
              schema:
                type: object
        '401':
          description: Admin token required
...
  /api/proof/benchmarks:
    get:
      summary: Public-safe provider benchmark matrix
      parameters:
        - in: query
          name: selectedOnly
          schema:
            type: boolean
            default: true
          description: When true, show only currently selected provider rows.
        - in: query
          name: provider
          schema:
            type: string
          description: Provider id or public label substring.
        - in: query
          name: providerId
          schema:
            type: string
        - in: query
          name: benchmarkId
          schema:
            type: string
            enum: [tiny_smoke, small_chat, summary_batch]
        - in: query
          name: status
          schema:
            type: string
            enum: [succeeded, failed, timed_out, not_allowed, untested]
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Provider benchmark matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkSummary'
    post:
      summary: Admin-only selected-provider benchmark run
      description: Runs one fixed benchmark definition through the provider-job proof adapter. The request accepts provider id, benchmark id, and prototype adapter mode only; prompt and output text are never accepted or stored.
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [providerId]
              properties:
                providerId:
                  type: string
                benchmarkId:
                  type: string
                  enum: [tiny_smoke, small_chat, summary_batch]
                adapterMode:
                  type: string
                  enum: [mock_success, mock_failure, mock_timeout, provider_http]
      responses:
        '200':
          description: Benchmark run completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  run:
                    $ref: '#/components/schemas/BenchmarkRun'
                  receipt:
                    $ref: '#/components/schemas/ProviderJobReceipt'
        '400':
          description: Invalid benchmark request
        '401':
          description: Admin token required
        '403':
          description: Provider is not selected for the benchmark workload
...
  /api/proof/payouts/batches:
    get:
      summary: Admin-only payout batch list
      security:
        - adminToken: []
      responses:
        '200':
          description: Payout batch list
          content:
            application/json:
              schema:
                type: object
                properties:
                  object:
                    type: string
                    example: list
                  dataState:
                    type: string
                    enum: [live, snapshot, sample, unavailable]
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PayoutBatch'
    post:
      summary: Admin-only payout review batch creation
      security:
        - adminToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                eventIds:
                  type: array
                  items:
                    type: string
                states:
                  type: array
                  items:
                    type: string
                    enum: [accrued, review, approved, paid, disputed, voided]
                operatorOwner:
                  type: string
                reason:
                  type: string
      responses:
        '201':
          description: Payout batch created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  batch:
                    $ref: '#/components/schemas/PayoutBatch'
...
  /api/staking/wallet-intents:
    get:
      summary: Public-safe signed OCEAN wallet intent list
      description: Lists wallet lock intents with hashed addresses and signatures only. These intents do not issue credits until a lock contract or operator verification flow confirms them.
      responses:
        '200':
          description: Public wallet intent summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletIntentSummary'
    post:
      summary: Record a signed OCEAN lock intent
      description: Browser-wallet prototype for holders to sign OCEAN lock intent. The route stores public-safe hashes and estimated credits, but does not stake tokens or issue Fish Credits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletIntentRequest'
      responses:
        '201':
          description: Wallet intent recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  intent:
                    $ref: '#/components/schemas/PublicWalletIntent'
                  notice:
                    type: string
        '400':
          description: Invalid signed wallet intent
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminToken:
      type: apiKey
      in: header
      name: x-fish-admin-token
  schemas:
...
    ProviderPilotRegistry:
      type: object
      properties:
        dataState:
          type: string
          enum: [live, snapshot, sample, unavailable]
        lastUpdated:
          type: string
        counts:
          type: object
          properties:
            applications:
              type: integer
            applied:
              type: integer
            allowed:
              type: integer
            fishReady:
              type: integer
            paused:
              type: integer
            rejected:
              type: integer
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderPilotProfile'
        allowlist:
          type: array
          items:
            $ref: '#/components/schemas/ProviderAllowlistEntry'
        allowedFor:
          nullable: true
          type: object
        warnings:
          type: array
          items:
            type: string
...
    ProviderJobReceipt:
      type: object
      properties:
        receiptVersion:
          type: integer
        receiptType:
          type: string
          example: provider_job_receipt
        receiptId:
          type: string
        jobId:
          type: string
        providerJobId:
          type: string
          nullable: true
        providerId:
          type: string
        providerLabel:
          type: string
        model:
          type: string
        workloadType:
          type: string
        backend:
          type: string
          enum: [ocean_provider]
        status:
          type: string
          enum: [succeeded, failed, timed_out, not_allowed]
        sourceState:
          type: string
          enum: [live, snapshot, sample, unavailable]
        visibility:
          type: string
        createdAt:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
        usage:
          type: object
        cost:
          type: object
        hashes:
          type: object
          properties:
            inputHash:
              type: string
            outputHash:
              type: string
              nullable: true
            canonicalReceiptHash:
              type: string
        signer:
          type: object
          properties:
            keyId:
              type: string
            algorithm:
              type: string
              enum: [none, ed25519]
            publicKeyPem:
              type: string
              nullable: true
        signatureStatus:
          type: string
          enum: [not_required, valid, invalid, missing]
        signature:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
...
    BenchmarkRun:
      type: object
      properties:
        benchmarkRunVersion:
          type: integer
          enum: [1]
        benchmarkRunId:
          type: string
        benchmarkId:
          type: string
          enum: [tiny_smoke, small_chat, summary_batch]
        providerId:
          type: string
        providerLabel:
          type: string
        modelClass:
          type: string
        workloadType:
          type: string
        model:
          type: string
        inputSizeBucket:
          type: string
        outputSizeBucket:
          type: string
        batchSize:
          type: integer
        adapterVersion:
          type: string
        status:
          type: string
          enum: [succeeded, failed, timed_out, not_allowed]
        sourceState:
          type: string
          enum: [live, snapshot, sample, unavailable]
        visibility:
          type: string
        receiptId:
          type: string
        receiptSignatureStatus:
          type: string
          enum: [not_required, valid, invalid, missing]
        jobId:
          type: string
        sampleSize:
          type: integer
        queueSeconds:
          type: number
          nullable: true
        runtimeSeconds:
          type: number
        gpuSeconds:
          type: number
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        tokensPerSecond:
          type: number
        providerCostUsd:
          type: number
        costPer1kTokensUsd:
          type: number
          nullable: true
        createdAt:
          type: string
...
    BenchmarkSummary:
      type: object
      properties:
        dataState:
          type: string
          enum: [live, snapshot, sample, unavailable]
        lastUpdated:
          type: string
        filters:
          type: object
        definitions:
          type: array
          items:
            type: object
        runs:
          type: array
          items:
            $ref: '#/components/schemas/BenchmarkRun'
        matrix:
          type: array
          items:
            $ref: '#/components/schemas/BenchmarkMatrixRow'
        report:
          $ref: '#/components/schemas/BenchmarkReport'
        totals:
          type: object
          properties:
            selectedProviders:
              type: integer
            benchmarkRuns:
              type: integer
            successfulRuns:
              type: integer
            failedRuns:
              type: integer
            timedOutRuns:
              type: integer
            untestedCells:
              type: integer
            passRate:
              type: number
              nullable: true
        warnings:
          type: array
          items:
            type: string
...
    PayoutBatch:
      type: object
      properties:
        payoutBatchId:
          type: string
        state:
          type: string
          enum: [review, approved, paid, voided]
        eventIds:
          type: array
          items:
            type: string
        providerCount:
          type: integer
        eventCount:
          type: integer
        amountUsd:
          type: number
        currency:
          type: string
        sourceState:
          type: string
          enum: [live, snapshot, sample, unavailable]
        visibility:
          type: string
        createdAt:
          type: string
        createdBy:
          type: string
        reason:
          type: string
        approvedAt:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        transactionRefs:
          type: array
          items:
            type: string
...
    CapacitySettlementRequest:
      type: object
      additionalProperties: false
      required: [grossUsdcAmount]
      properties:
        grossUsdcAmount:
          type: number
          minimum: 0.000001
        netUsdcAmount:
          type: number
          minimum: 0
        operatorFeeUsdc:
          type: number
          minimum: 0
        settlementSource:
          type: string
          enum: [stripe, usdc_checkout, api_subscription, operator_adjustment]
          default: operator_adjustment
        transactionHash:
          type: string
          nullable: true
          pattern: '^0x[a-fA-F0-9]{64}$'
        capacityPoolAddress:
          type: string
          nullable: true
          pattern: '^0x[a-fA-F0-9]{40}$'
        paidDemandRef:
          type: string
          nullable: true
        idempotencyKey:
          type: string
          nullable: true
          minLength: 8
          description: Required when submitOnchain is true.
        submitOnchain:
          type: boolean
          default: false
        occurredAt:
          type: string
        note:
          type: string
          nullable: true
    CapacitySettlement:
      type: object
      properties:
        settlementVersion:
          type: integer
          enum: [1]
        settlementId:
          type: string
        settlementType:
          type: string
          enum: [fish_capacity_pool_settlement]
        grossUsdcAmount:
          type: number
        netUsdcAmount:
          type: number
        operatorFeeUsdc:
          type: number
        settlementSource:
          type: string
          enum: [stripe, usdc_checkout, api_subscription, operator_adjustment]
        transactionHash:
          type: string
          nullable: true
        transactionHashPrefix:
          type: string
          nullable: true
        approveTransactionHash:
          type: string
          nullable: true
        submitState:
          type: string
          enum: [not_submitted, submitted]
        onchainSubmission:
          type: object
          nullable: true
          properties:
            chainId:
              type: integer
            operatorAddressPrefix:
              type: string
            grossAmountRaw:
              type: string
            operatorFeeBps:
              type: integer
            recordPaidUsageBlockNumber:
              type: integer
              nullable: true
            confirmations:
              type: integer
        capacityPoolAddress:
          type: string
          nullable: true
        capacityPoolAddressPrefix:
          type: string
          nullable: true
        paidDemandHash:
          type: string
          nullable: true
        idempotencyHash:
          type: string
          nullable: true
        sourceState:
          type: string
          enum: [live, snapshot]
        visibility:
          type: string
          enum: [public]
        occurredAt:
          type: string
        createdAt:
          type: string
    CapacitySettlementSummary:
      type: object
      properties:
        dataState:
          type: string
          enum: [live, snapshot, sample, unavailable]
        lastUpdated:
          type: string
        totals:
          type: object
          properties:
            settlements:
              type: integer
            grossUsdcAmount:
              type: number
            netUsdcAmount:
              type: number
            operatorFeeUsdc:
              type: number
            onchainSubmittedSettlements:
              type: integer
        configuredCapacityPoolAddress:
          type: string
          nullable: true
        settlements:
          type: array
          items:
            $ref: '#/components/schemas/CapacitySettlement'
        warnings:
          type: array
          items:
            type: string
...
    FishChatRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: string
          default: fish-demo-chat
          description: Use `fish-demo-chat` for the generic route, or a dish alias such as `fish-ask`, `fish-code`, `fish-docs`, `fish-repo`, `fish-eval`, `fish-data`, `fish-ocean-helper`, `fish-clear-broth`, or `fish-proposal`.
        messages:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant, tool]
              content:
                oneOf:
                  - type: string
                  - type: array
                    items: {}
        stream:
          type: boolean
          default: false
        temperature:
          type: number
          minimum: 0
          maximum: 2
        max_tokens:
          type: integer
          minimum: 1
          maximum: 4096
        metadata:
          type: object
          additionalProperties: true
          properties:
            fish_feature:
              type: string
              enum: [ask, code, explain, docs, repo, eval, data, images, proposal, ocean, api]
            fish_dish:
              type: string
            requestedPrivacyMode:
              type: string
              enum: [local_demo, external_policy, ocean_demo_policy, selected_ocean_policy, hash_only_batch, ocean_hardened, ocean_tee, e2ee_to_tee]
              description: Optional requested privacy mode. Fish rejects unavailable modes unless `allowPrivacyDowngrade` is true.
            allowPrivacyDowngrade:
              type: boolean
              default: false
...
    FishChatCompletion:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        fish:
          type: object
          properties:
            route:
              type: string
              enum: [mock, ocean-demo-vllm, ocean-provider, external-fallback]
            requestedRoute:
              type: string
              nullable: true
              enum: [mock, ocean-demo-vllm, ocean-provider, external-fallback]
            fallbackFrom:
              type: string
              nullable: true
              enum: [mock, ocean-demo-vllm, ocean-provider, external-fallback]
            fallbackReason:
              type: string
              nullable: true
            feature:
              type: string
              nullable: true
            featureLabel:
              type: string
            dishId:
              type: string
              nullable: true
            accessMode:
              type: string
              nullable: true
              enum: [guest, key]
            costState:
              type: string
              enum: [prototype_estimate, provider_verified, fallback_verified]
            receiptId:
              type: string
            status:
              type: string
            routeLabel:
              type: string
            providerId:
              type: string
              nullable: true
            runnerReceiptHash:
              type: string
              nullable: true
            runnerSignatureState:
              type: string
              nullable: true
              enum: [unsigned, signed, verified, invalid, missing]
            runnerId:
              type: string
              nullable: true
            latencyMs:
              type: integer
            dailyBudgetUsd:
              type: number
            dailyBudgetRemainingUsd:
              type: number
            estimatedProviderCostUsd:
              type: number
            quotaRemaining:
              type: integer
            monthlyRequestRemaining:
              type: integer
              nullable: true
            monthlyRequestResetAt:
              type: string
              nullable: true
            rateLimitRemaining:
              type: integer
              nullable: true
            rateLimitResetAt:
              type: string
              nullable: true
            creditReserveId:
              type: string
            creditsReserved:
              type: integer
            creditsReleased:
              type: integer
            creditsSpent:
              type: integer
            creditsRemaining:
              type: integer
            userChargeUsd:
              type: number
            providerCostUsd:
              type: number
            grossMarginUsd:
              type: number
            knowledgeSources:
              type: array
              items:
                $ref: '#/components/schemas/FishKnowledgeSource'
            privacy:
              $ref: '#/components/schemas/FishUsagePrivacy'
            batchReceiptId:
              type: string
            batchJobId:
              type: string
            batchSourceState:
              type: string
              enum: [live, snapshot, sample, unavailable]
            batchAdapterMode:
              type: string
              enum: [sample_success, sample_failure, ocean_http]
            batchOutputRef:
              type: string
              nullable: true
            inputRef:
              type: string
            storesPromptOutputText:
              type: boolean
              enum: [false]
            storesOutputText:
              type: boolean
              enum: [false]
...
    FishKnowledgeSource:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        source:
          type: string
...